refactor(otlp-transformer): re-struture package to prepare for separate entrypoints
