+++ experimental/packages/otlp-transformer/src/common/protobuf/protobuf-export-type.ts
