+++ experimental/packages/otlp-transformer/src/i-serializer.ts
