+++ experimental/packages/otlp-transformer/src/logs/json/logs.ts
@@ -1,0 +1,40 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { ISerializer } from '../../i-serializer';
+import { ReadableLogRecord } from '@opentelemetry/sdk-logs';
+import { createExportLogsServiceRequest } from '../internal';
+import { IExportLogsServiceResponse } from '../export-response';
+
+/*
+ * @experimental this serializer may receive breaking changes in minor versions, pin this package's version when using this constant
+ */
+export const JsonLogsSerializer: ISerializer<
+  ReadableLogRecord[],
+  IExportLogsServiceResponse
+> = {
+  serializeRequest: (arg: ReadableLogRecord[]) => {
+    const request = createExportLogsServiceRequest(arg, {
+      useHex: true,
+      useLongBits: false,
+    });
+    const encoder = new TextEncoder();
+    return encoder.encode(JSON.stringify(request));
+  },
+  deserializeResponse: (arg: Uint8Array) => {
+    const decoder = new TextDecoder();
+    return JSON.parse(decoder.decode(arg)) as IExportLogsServiceResponse;
+  },
+};
