+++ experimental/packages/otlp-transformer/src/trace/export-response.ts
@@ -14,8 +14,15 @@
  * limitations under the License.
  */
 
-// IMPORTANT: exports added here are public
-export {
-  IExportTracePartialSuccess,
-  IExportTraceServiceResponse,
-} from './export-response';
+export interface IExportTraceServiceResponse {
+  /** ExportTraceServiceResponse partialSuccess */
+  partialSuccess?: IExportTracePartialSuccess;
+}
+
+export interface IExportTracePartialSuccess {
+  /** ExportLogsServiceResponse rejectedLogRecords */
+  rejectedSpans?: number;
+
+  /** ExportLogsServiceResponse errorMessage */
+  errorMessage?: string;
+}
