+++ experimental/packages/otlp-transformer/src/trace/json/trace.ts
@@ -1,0 +1,37 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { ISerializer } from '../../i-serializer';
+import { ReadableSpan } from '@opentelemetry/sdk-trace-base';
+import { IExportTraceServiceResponse } from '../export-response';
+import { createExportTraceServiceRequest } from '../internal';
+
+export const JsonTraceSerializer: ISerializer<
+  ReadableSpan[],
+  IExportTraceServiceResponse
+> = {
+  serializeRequest: (arg: ReadableSpan[]) => {
+    const request = createExportTraceServiceRequest(arg, {
+      useHex: true,
+      useLongBits: false,
+    });
+    const encoder = new TextEncoder();
+    return encoder.encode(JSON.stringify(request));
+  },
+  deserializeResponse: (arg: Uint8Array) => {
+    const decoder = new TextDecoder();
+    return JSON.parse(decoder.decode(arg)) as IExportTraceServiceResponse;
+  },
+};
