+++ experimental/packages/otlp-transformer/src/trace/protobuf/index.ts
@@ -15,7 +15,4 @@
  */
 
 // IMPORTANT: exports added here are public
-export {
-  IExportTracePartialSuccess,
-  IExportTraceServiceResponse,
-} from './export-response';
+export { ProtobufTraceSerializer } from './trace';
