+++ experimental/packages/otlp-transformer/test/common.test.ts
@@ -17,7 +17,7 @@
 import { hexToBinary } from '@opentelemetry/core';
 import { toAnyValue } from '../src/common/internal';
 import * as assert from 'assert';
-import { getOtlpEncoder } from '../src/common';
+import { getOtlpEncoder } from '../src/common/utils';
 
 const traceId = 'abcdef01234567890000000000000000';
 const spanId = '12341234abcdabcd';
